% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-GetPaths.R
\name{get_microbs_ddPCR_rawDataPath}
\alias{get_microbs_ddPCR_rawDataPath}
\title{get ddPCR raw data path}
\usage{
get_microbs_ddPCR_rawDataPath()
}
\description{
If it returns NULL, it means that the path connector has not been set yet.
Call set_microbs_ddPCR_rawDataPath() to set the path connector.
Add a string to set custom path connector.
Only use the directory name, not the full path.
This function allows you to get the path connector between global directory & the data directories.

file ./microbs.lu/R/02_GetPaths.R
}
\examples{
# Example usage
result <- get_microbs_ddPCR_rawDataPath()

result

}
