% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_SetPaths.R
\name{set_microbs_connector_dir}
\alias{set_microbs_connector_dir}
\title{Set path connector in which the data is saved}
\usage{
set_microbs_connector_dir(path = "Data_Treatment")
}
\arguments{
\item{path}{A character string representing the path connector.
The default path is "Data_Treatment".}
}
\description{
Within theteam there is a global directory, and within this directory there is another directory called "Data_Treatment".
And finally the data directories are within this directory.
This function allows you to set the path connector bwtween global directory & the data directories.

This is not really needed if you are using absolute paths (Complete path for example \verb{/home/user/myWorkingDir/...}).
file ./microbs.lu/R/02_SetPaths.R
}
\examples{
# Example usage
result <- set_microbs_connector_dir(path="Data_Treatment")

result

}
