% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_SetPaths.R
\name{set_microbs_qPCR_rawDataPath}
\alias{set_microbs_qPCR_rawDataPath}
\title{Set dPCR Raw Data Path}
\usage{
set_microbs_qPCR_rawDataPath(path = "0_raw_data_qPCR", relative = FALSE)
}
\arguments{
\item{path}{A character string representing the path to be set as the working directory.
The default path is "0_raw_data_qPCR".}

\item{relative}{A Boolean to use relative path or not. Default is False.}
}
\description{
The raw data from the qPCR tests are to be stored in the "0_raw_data_qPCR" directory.
It is assumed that the global directory is set using the set_microbs_wdirectory function.
Within this directory there is another director called "Data_Treatment", set using the function set_microbs_connector_dir.
This is just because the direectory structure within the team is like this.
It is not necessary to have a connector (in-between) directory.
This function allows you to set the path for the raw ddPCR data.
file ./microbs.lu/R/02_SetPaths.R
}
\examples{
# Example usage
set_microbs_wdirectory() # to set default working directory
set_microbs_connector_dir() # to set default connector directory
path <- "0_raw_data_qPCR"
result <- set_microbs_qPCR_rawDataPath(path)
result

path <- "D:/03_Workspace/01_R_Package/microbs_lu_dummy_data/Data_Treatment/0_raw_data_qPCR"
result <- set_microbs_qPCR_rawDataPath(path)
result

}
