% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_SetPaths.R
\name{set_microbs_wdirectory}
\alias{set_microbs_wdirectory}
\title{Set the Microbs Working Directory}
\usage{
set_microbs_wdirectory(
  path = "L:/Units & Programmes/BIOTECH/ENVMICRO/_Common/Projects/SUPERVIR/11-Results"
)
}
\arguments{
\item{path}{A character string representing the path to be set as the working directory.
The default path is "L:/Units & Programmes/BIOTECH/ENVMICRO/_Common/Projects/SUPERVIR/11-Results".
file ./microbs.lu/R/02_SetPaths.R}
}
\description{
This function allows you to set the working directory for the package. This funciton uses \code{setwd()}
to set the working directory. A default path is already provided for LIST.

This script sets the paths for various directories used in the microbs.lu package.
}
\details{
The working directory is the directory where the package will look for data files.
If you want to use a custom path, make sure you have the folder structure & respective data in place.
\strong{Note:} Use '/' instead of '\' in the path.

The following directory structure is required for the package, these directories are located inside :
.
├── Data_Treatment
│   ├── 00_flux_data
│   ├── 00_standard_curve
│   ├── 0_raw_data_ddPCR
│   ├── 0_raw_data_qPCR
│   ├── 1_loaded_data
│   ├── 2_calc_data
│   ├── 3_created_data
│   └── 4_data_4_dashboard
├── microbs_runtime.R
}
