% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_microbs_path_builder}
\alias{utils_microbs_path_builder}
\title{Builds the complete path}
\usage{
utils_microbs_path_builder(global_dir_path, connector, local_dir_path)
}
\arguments{
\item{global_dir_path}{A string that is the global directory path.}

\item{connector}{A string is the connector between the global directory path & the local directory path.}

\item{local_dir_path}{A string, that is usually the data directory paths, example \verb{0_raw_data_qPCR}.}
}
\value{
A string that is a complete path, rather than a relavtive path known in the repo.
}
\description{
This is an example function that demonstrates the structure of an R function.
}
\examples{
# Example usage
global_dir_path <- "L:/Units & Programmes/BIOTECH/ENVMICRO/_Common/Projects/SUPERVIR/11-Results"
connector <- "Data_Treatment"
local_dir_path <- "0_raw_data_qPCR"
result <- utils_microbs_path_builder(global_dir_path, connector, local_dir_path)
print(result) 
}
